+++ src/constants.js
@@ -9,4 +9,5 @@ export const FLAG_DUNGEON_VERSION = "dungeonVersion";
 export const SETTING_3DCANVAS_ENABLED = "3DCanvasEnabled";
 export const SETTING_CUSTOM_THEMES = "customThemes";
 export const SETTING_RELEASE_NOTES_VERSION = "releaseNotesVersion";
+export const SETTING_SNAP_TO_GRID = "snapToGrid";
 export const SETTING_THEME_PAINTER_THEME = "themePainterTheme";