+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.14.0
 
-- Add support for 3D Canvas module wall flags.
+- Add snap to grid, with module setting to turn on/off. Hold shift to disable.
+- Add support for 3D Canvas module wall flags, with module setting to turn on/off.
 - Add "Save to Scene Background" button, useful for setting a 3D Canvas floor.
 - Change renderer to use Pixi.js polygon texture fill for floor and wall textures, rather than manually creating Sprites. This should allow small (e.g., 100x100px) textures to be performant.
 - Only show delete-all button to GM.